Add new GitHub Actions workflow for .NET project
Introduce a new GitHub Actions workflow configuration in `xUnitTests.yml` for building and testing the .NET project. The workflow triggers on pushes and pull requests to the "master" branch and runs on `ubuntu-latest`. It includes steps to check out the repository, set up .NET 8.0.x, restore dependencies, build the project, and run tests.

Update `ProxySharp.csproj` to remove the reference to the old workflow file `.github\workflows\dotnet.yml` and add a new item group to include the `.github\workflows\` folder.